Refactor frontend components: Extract business logic to custom hooks
