+++ web/src/pages/Dashboard/Dashboard.jsx
@@ -1,16 +1,5 @@
-import {
-  INDEX_SYMBOLS,
-  fallbackIndex,
-  getCurrentUser,
-  getIndices,
-  normalizeIndexSymbol,
-  getNews,
-} from './utils/api';
-import { fetchMarketStatus } from '@/lib/marketUtils';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useToast } from '@/components/ui/use-toast';
-import { getFlashWorkspace } from '../ChatAgent/utils/api';
 import { useNavigate } from 'react-router-dom';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription } from '../../components/ui/dialog';
 import { Input } from '../../components/ui/input';
@@ -29,58 +18,13 @@ import AddPortfolioHoldingDialog from './components/AddPortfolioHoldingDialog';
 import { useWatchlistData } from './hooks/useWatchlistData';
 import { usePortfolioData } from './hooks/usePortfolioData';
 import { useTickerNews } from './hooks/useTickerNews';
+import { useDashboardData } from './hooks/useDashboardData';
+import { useOnboarding, setOnboardingIgnoredFor24h } from './hooks/useOnboarding';
 import './Dashboard.css';
 
-
-const ONBOARDING_IGNORE_STORAGE_KEY = 'langalpha-onboarding-ignored-at';
-const ONBOARDING_IGNORE_MS = 24 * 60 * 60 * 1000; // 24 hours
-
-function isOnboardingIgnoredFor24h() {
-  try {
-    const stored = localStorage.getItem(ONBOARDING_IGNORE_STORAGE_KEY);
-    if (!stored) return false;
-    const timestamp = parseInt(stored, 10);
-    if (Number.isNaN(timestamp)) return false;
-    return Date.now() - timestamp < ONBOARDING_IGNORE_MS;
-  } catch {
-    return false;
-  }
-}
-
-function setOnboardingIgnoredFor24h() {
-  try {
-    localStorage.setItem(ONBOARDING_IGNORE_STORAGE_KEY, String(Date.now()));
-  } catch (e) {
-    console.warn('[Dashboard] Could not persist onboarding ignore', e);
-  }
-}
-
-// Module-level caches (survive navigation, clear on page refresh)
-let newsCache = null;          // { items }
-let indicesCache = null;       // [ index objects ]
-
-function formatRelativeTime(timestamp) {
-  if (!timestamp) return '';
-  const now = new Date();
-  const then = new Date(timestamp);
-  const diffMs = now - then;
-  const diffMin = Math.floor(diffMs / 60000);
-  if (diffMin < 1) return 'just now';
-  if (diffMin < 60) return `${diffMin} min ago`;
-  const diffHr = Math.floor(diffMin / 60);
-  if (diffHr < 24) return `${diffHr} hr${diffHr > 1 ? 's' : ''} ago`;
-  const diffDay = Math.floor(diffHr / 24);
-  return `${diffDay} day${diffDay > 1 ? 's' : ''} ago`;
-}
-
 function Dashboard() {
-  const { toast } = useToast();
   const navigate = useNavigate();
   const { t } = useTranslation();
-
-  // Onboarding check state
-  const [showOnboardingDialog, setShowOnboardingDialog] = useState(false);
-  const [isCreatingWorkspace, setIsCreatingWorkspace] = useState(false);
 
   // News modal state
   const [selectedNewsId, setSelectedNewsId] = useState(null);
@@ -88,152 +32,21 @@ function Dashboard() {
   // Insight modal state
   const [selectedMarketInsightId, setSelectedMarketInsightId] = useState(null);
 
-  const [indices, setIndices] = useState(() =>
-    indicesCache || INDEX_SYMBOLS.map((s) => fallbackIndex(normalizeIndexSymbol(s)))
-  );
-  const [indicesLoading, setIndicesLoading] = useState(!indicesCache);
+  const {
+    indices,
+    indicesLoading,
+    newsItems,
+    newsLoading,
+    marketStatus,
+  } = useDashboardData();
 
-  const [newsItems, setNewsItems] = useState(() => newsCache?.items || []);
-  const [newsLoading, setNewsLoading] = useState(!newsCache);
-
-  const fetchNews = useCallback(async () => {
-    setNewsLoading(true);
-    try {
-      const data = await getNews({ limit: 50 });
-      if (data.results && data.results.length > 0) {
-        const mapped = data.results.map((r) => ({
-          id: r.id,
-          title: r.title,
-          time: formatRelativeTime(r.published_at),
-          isHot: r.has_sentiment,
-          source: r.source?.name || '',
-          favicon: r.source?.favicon_url || null,
-          image: r.image_url || null,
-          tickers: r.tickers || [],
-        }));
-        setNewsItems(mapped);
-        newsCache = { items: mapped };
-      }
-    } catch {
-      // Keep existing items on error; empty array if first load
-    } finally {
-      setNewsLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!newsCache) fetchNews();
-  }, [fetchNews]);
-
-  const fetchIndices = useCallback(async () => {
-    if (!indicesCache) setIndicesLoading(true);
-    try {
-      const { indices: next } = await getIndices(INDEX_SYMBOLS);
-      setIndices(next);
-      indicesCache = next;
-    } catch (error) {
-      console.error('[Dashboard] Error fetching indices:', error?.message);
-      if (!indicesCache) {
-        setIndices(INDEX_SYMBOLS.map((s) => fallbackIndex(normalizeIndexSymbol(s))));
-      }
-    } finally {
-      setIndicesLoading(false);
-    }
-  }, []);
-
-  // Adaptive polling: 30s during market hours, 60s during extended/closed
-  const marketStatusRef = useRef(null);
-  const [marketStatus, setMarketStatus] = useState(null);
-  useEffect(() => {
-    const pollMarketStatus = () =>
-      fetchMarketStatus().then((s) => { marketStatusRef.current = s; setMarketStatus(s); }).catch(() => {});
-    pollMarketStatus();
-    const statusId = setInterval(pollMarketStatus, 60000);
-    return () => clearInterval(statusId);
-  }, []);
-
-  useEffect(() => {
-    fetchIndices();
-    let intervalId = null;
-    const schedule = () => {
-      const isMarketOpen = marketStatusRef.current?.market === 'open'
-        || (marketStatusRef.current && !marketStatusRef.current.afterHours && !marketStatusRef.current.earlyHours && marketStatusRef.current.market !== 'closed');
-      const delay = isMarketOpen ? 30000 : 60000;
-      intervalId = setTimeout(() => {
-        if (!document.hidden) fetchIndices();
-        schedule();
-      }, delay);
-    };
-    schedule();
-    return () => { if (intervalId) clearTimeout(intervalId); };
-  }, [fetchIndices]);
-
-  /**
-   * Check onboarding completion status on every Dashboard mount.
-   */
-  useEffect(() => {
-    const checkOnboarding = async () => {
-      try {
-        const userData = await getCurrentUser();
-        const onboardingCompleted = userData?.user?.onboarding_completed;
-
-        if (onboardingCompleted === true) {
-          setShowOnboardingDialog(false);
-          return;
-        }
-        if (onboardingCompleted === false && !isOnboardingIgnoredFor24h()) {
-          setShowOnboardingDialog(true);
-        }
-      } catch (error) {
-        console.error('[Dashboard] Error checking onboarding status:', error);
-      }
-    };
-
-    checkOnboarding();
-  }, []);
-
-  const navigateToOnboarding = useCallback(async () => {
-    setIsCreatingWorkspace(true);
-    try {
-      const flashWs = await getFlashWorkspace();
-      navigate(`/chat/${flashWs.workspace_id}/__default__`, {
-        state: {
-          isOnboarding: true,
-          agentMode: 'flash',
-          workspaceStatus: 'flash',
-        },
-      });
-    } catch (error) {
-      console.error('Error setting up onboarding:', error);
-      toast({
-        variant: 'destructive',
-        title: t('common.error'),
-        description: t('dashboard.failedOnboarding'),
-      });
-    } finally {
-      setIsCreatingWorkspace(false);
-    }
-  }, [navigate, toast]);
-
-  const navigateToModifyPreferences = useCallback(async () => {
-    try {
-      const flashWs = await getFlashWorkspace();
-      navigate(`/chat/${flashWs.workspace_id}/__default__`, {
-        state: {
-          isModifyingPreferences: true,
-          agentMode: 'flash',
-          workspaceStatus: 'flash',
-        },
-      });
-    } catch (error) {
-      console.error('Error navigating to modify preferences:', error);
-      toast({
-        variant: 'destructive',
-        title: t('common.error'),
-        description: t('dashboard.failedPrefUpdate'),
-      });
-    }
-  }, [navigate, toast]);
+  const {
+    showOnboardingDialog,
+    setShowOnboardingDialog,
+    isCreatingWorkspace,
+    navigateToOnboarding,
+    navigateToModifyPreferences
+  } = useOnboarding();
 
   const watchlist = useWatchlistData();
   const portfolio = usePortfolioData();
